requirements.txt
arrow==0.15.5
asgiref==3.2.7
binaryornot==0.4.4
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
Django==3.0.5
djangorestframework==3.11.0
idna==2.9
Jinja2==2.11.2
jinja2-time==0.2.0
MarkupSafe==1.1.1
poyo==0.5.0
python-dateutil==2.8.1
python-slugify==4.0.0
pytz==2020.1
requests==2.23.0
six==1.14.0
sqlparse==0.3.1
text-unidecode==1.3
urllib3==1.25.9
